+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 Here, you will find all notabled changes that have been made to this project.
 
+## 1.0.2 - 2024-01-05
+
+### Added
+- Added weekly downloads to "packageDownloads" and "allDownloads"
+- Added more keywords to package.json
+
+### Changed
+
+### Fixed
+
 ## 1.0.2 - 2024-01-04
 
 ### Added